Consistent formatting
In fb84d67, we made some changes to fix formatting, and
matched some of legacy pambase's formatting even where it was
inconsistent to be sure that things would work.

With the fixed line endings in that commit (\n, not \r\n),
we're all good, so let's be more consistent with tabs.

Fixes: fb84d6762243b9c488226a07e8d14b4041cc2944
Signed-off-by: Sam James <sam@gentoo.org>